refactor: :truck: move interface to interface folder
